Add optional request/response logging
This commit adds an optional filter that logs requests and responses.

This filter can only be enabled in ‘test’ environments. To support this, a discrete spring profile is defined for each environment and an `EnvironmentService` has been added to determine the type of environment the code is currently running in.

The ‘test’ profile has been renamed to ‘integration-test’, in anticipation for a standalone ‘test’ environment being added

An example of how requests appear:

```
2025-08-29 11:54:50.737 DEBUG 15098 --- [o-auto-1-exec-4] ion$CommunityPaybackRequestLoggingFilter : Before request [POST /example] |
2025-08-29 11:54:50.741 DEBUG 15098 --- [o-auto-1-exec-4] ion$CommunityPaybackRequestLoggingFilter : Request data: POST /example, payload={"apiName":"test-api"}] |
```