Add cp-stack tool (#19)
The cp-stack tool provides a simple way to run the components of the community-payback stack on developers machines.

Initially this delegates to upstream services hosted in the dev environment. This can be changed to use locally deployed services if required in the future. For notes on the pros/cons of this approach, see https://dsdmoj.atlassian.net/browse/CPB-51.