CPB-84 Add upstream client for providers using wiremock response (#43)
* Use WireMock Spring Support

This commit adds the wiremock spring boot library and uses it in integration tests via the `@EnableWireMock` annotation.

This allows us to simplify the `HmppsAuthMockServer` configuration as we delegate managing the wiremock server during tests to wiremock itself.

* Add upstream client using wiremock response

This commit adds a client to call an upstream `community-payback-and-delius` API. On deployment this client is currently configured to loopback to the API calling the `/mocks/community-payback-and-delius` endpoint.

This commit also adds configuration to `cp-stack` to route the upstream call via the wiremock instance hosted in `cp-stack`, demonstrating how we can intercept upstream calls in local dev to return canned responses.