Add offender service and upstream endpoints
This commit adds an `OffenderService` that checks if the calling user has limited access to the corresponding offender(s) and if they do, returns a limited view of the offender.

To support this, upstream client calls have also been added, assuming the upstream service will use the same pattern used by the approved-premises-api project.