CPB-85: Automate API type generation in UI repository (#73)
* Add generate types workflow

This will trigger a workflow the UI repo, and is based on a set of workflows in the `approved-premesis-api` repository.

Since switching from the `tibdex/github-app-token@v1` action (deprecated), to GitHub's `actions/create-github-app-token@v2` we need to add the permission we want the token to have, and the reposiories the permission should be for, in order for a successful `/dispatches` post request.

Without this we get the error `RequestError [HttpError]: Resource not accessible by integration`.

* Add genereate types workflow to pipeline steps