Publish Domain Event on Appointment Update
This commit adds a `DomainEventService` to build a thin domain event for a given Domain Event Type, and delegates to the `DomainEventPublisher` to publish the domain event. If there is a transaction the publisher will wait until the transaction commits before publishing the domain event.

The `DomainEventService` is then used to raise a domain event whenever an appointment is updated.

The endpoint used to retrieve domain event details will be added in a susbsequent commit.