Remove duplication from Open Api Annotations
Before this commit we were duplicating some elements of the API already defined by code or spring annotations in the Open API Annotations. This made it possible to declare one reponse type in the function but document a different response type in Open API.

This commit removes all redundant open api annotations, using code or spring annotations to determine elements of the documentation instead.