Update upstream get appointment response
The response has been modified to be more structured. This commit updates the sample responses and updates the mapping code to use these new structures. It also adds pickup location information which is now available.

Whilst case (offender) information has been added into the appointment response, we’re still fetching this via a separate API call. In a subsequent commit the case information in the appointment will be used instead.