Add supervisor endpoint and discrete roles
This commit adds a supervisor get sessions endpoint, and re-organises the controllers to indicate which UI which uses which endpoints.

In a subsequent commit the path used for all admin ui requests will be prefixed with `/admin`, which will simplify various configuration.

Some additional fields are required on the supervisor get sessions endpoint. These will also be added via a subsequent commit.