Remove duplicate entity refs in appt outcome
Remove dual mapping in the `AppointmentOutcomeEntity` for linked entities. As part of the move to an appointment update entity, we need to resolve these entities when building the request, these can’t currnetly be resolved on the entity model in the service function that persists the entiy, because it’s only writing IDs.

Given this, it’s going to be simplier to remove the dual mapping