Ensure appointment update validation returns 400
Before this commit the `AppointmentOutcomeValidationService` was using Kotlin’s `requireNotNull` and `require` functions to validate the `UpdateAppointmentOutcomeDto`. If validation failed these functions would return an `IllegalArgumentException` which would be returned as a HTTP 500 error to the caller.

This commit updates that service to instead use a custom `validateNotNull` function that will return a `BadRequestException`, for which we have mapping defined to instead return a HTTP 400 error.