Only populate appt.outcomeData for attended appts
Before this commit the `AppointmentDto.attendanceData` value was always populated when retrieving an appointment, regardless of whether attendance was recorded. This commit changes that behaviour to only return a value if an appointment outcome has been recorded against the appointment that indicates the appointment was attended.

This then allows us to make some fields on `AttendanceDataDto` mandatory