Outcome update fails if contact outcome is changed
The code before this commit would rely on `CompareToBuilder.reflectionCompare` to check linked JPA entities on the `AppointmentOutcomeEntity` for changes. This fails with a `ClassCastException` because the `reflectionCompare` struggles when deep comparing objects in this scenario.

This commit updates the `isLogicallyIdentical` fun to check linked JPA entities for changes directly, ensuring the `reflectionCompare` function only checks primitive values for changes. This commit also moves the `isLogicallyIdentical` function onto the `AppointmentOutcomeEntity` to better indicate that the function should be reviewed and updated when new JPA relationships are added.