Use new path for upstream session search endpoint
This commit updates the code to use a new path for the upstream endpoint to search for sessions, aligning with the code implemented in `community-payback-and-delius`. We continue to use a mocked endpoint until the get session upstream implementation is available.

This commit also adds a new UI facing session search API `/admin/providers/{providerCode}/teams/{teamCode}/sessions` that will capture the provider ID along with the existing team code, as this is required by the upstream API. The existing session search API has been deprecated in favour of the new API.

For now both APIs will continue to work as before.