Update upstream get appointments URL
This commit updates the upstream URL to align with the implemented version.

The existing UI facing API does not capture project, so we’re populating that value with ‘UNKNOWN’. A subsequent commit will add a new UI facing API that will require project, and the existing API will be deprecated.