Use custom type for API hour minute durations
We use the format `HH:MM` to express penalty time in both downstream and upstream JSON. Before this commit this value was mapped to a `LocalTime`, which is incorrect because this value represents a duration, not a time.

This commit adds a new `HourMinuteDuration` type that wraps a Java `Duration` for use for the aforementioned penalty time fields. This commit also adds and configures jackson configuration to handle serialization and deserialization.