Add script to clear form data
Add script to clear form data

There is a need for a method to quickly clear form data when testing.

This commit adds a test-support controller with an endpoint to clear form data.

Access to this endpoint is limited in several ways:

* The `test-support/*` controller is enabled by configuration, and can only be enabled in test environments
* The `test-support/*` endpoints are explicitly excluded from the ingress
* Calls not originating from localhost will be rejected

Therefore, it can only be called when the API is deployed locally, or the caller is connected to a k8s pod. A script has been added to support this.

To allow the endpoint to be called without an auth token a custom `ResourceServerConfiguration` has been added, matching the version defined in the [hmpps-sqs test app](https://github.com/ministryofjustice/hmpps-spring-boot-sqs/tree/main/test-app). This conveniently also allows access to the `/queue-admin/retry-all-dlqs` endpoint without a token, which will be required when working with the community campus message queue.