Remove contact outcomes from some mock data
We expect that in the majority of cases, appointments will not already have an
outcome for our users. This is particularly true in the Supervisor UI.

This sets the majority of contact outcomes (and attendance data) to null for
more realistic cases. Left some so we can see the handling of 'Not expected'
cases.