Replace UI project setup scripts with `script/bootstrap`
We have been changing the npm scripts used for setup fairly often recently at
the direction of the hmpps development team. This replaces the npm script used
in the tiltfile with the `script/bootstrap` command in the respective UI
projects, so that we only need to update these commands from the UI projects in
the future.